idGenerator added to serviceInformation

+++ ums-api/src/main/java/org/ums/domain/model/immutable/registrar/ServiceInformationDetail.java
@@ -20,5 +20,5 @@ public interface ServiceInformationDetail extends Serializable, EditType<Mutable
 
   Date getEndDate();
 
-  int getServiceId();
+  Long getServiceId();
 }
