+++ ums-core/src/main/java/org/ums/cache/registrar/ServiceInformationDetailCache.java
@@ -29,7 +29,7 @@ public class ServiceInformationDetailCache extends
   }
 
   @Override
-  public List<ServiceInformationDetail> getServiceInformationDetail(int pServiceId) {
+  public List<ServiceInformationDetail> getServiceInformationDetail(Long pServiceId) {
     return getManager().getServiceInformationDetail(pServiceId);
   }
